jddcnn: maybe this training works out?

+++ src/rawnind/libs/abstract_trainer.py
@@ -1730,8 +1730,8 @@ class BayerImageToImageNNTraining(ImageToImageNNTraining, BayerImageToImageNN):
 
         # unfortunately this doesn't always work with f16, gradients become nan.
         # might need to run a few iterations with float32 and then reduce precision as a fine tuning pass
-        # with torch.autocast(device_type="cuda",dtype=torch.float16):
-        with torch.autocast(device_type="cuda",dtype=torch.float32):
+        # with torch.autocast(device_type="cuda",dtype=torch.float32):
+        with torch.autocast(device_type="cuda",dtype=torch.float16):
             model_output = self.model(batch["y_crops"])
             if isinstance(self, DenoiseCompressTraining):
                 reconstructed_image, bpp = (
